Splitted legacy code into modules, improved reusability.
