+++ core/model.py
@@ -1,0 +1,33 @@
+import torch
+from transformers import AutoModelForCausalLM, AutoTokenizer, pipeline
+from langchain.llms import HuggingFacePipeline
+
+def prepare_pipeline(model, tokenizer):
+    return pipeline(
+        "text-generation",
+        model=model,
+        tokenizer=tokenizer,
+        max_new_tokens=200,
+        temperature=0.2,
+        top_p=0.9,
+        eos_token_id=tokenizer.eos_token_id,
+        pad_token_id=tokenizer.eos_token_id,
+        do_sample=True,
+        return_full_text=False,
+    )
+
+def load_llm(model_name="meta-llama/Llama-3.2-3B-Instruct"):
+    """Load LLaMA model wrapped in HuggingFacePipeline."""
+    print(f"Loading model {model_name}...")
+    tokenizer = AutoTokenizer.from_pretrained(model_name)
+    model = AutoModelForCausalLM.from_pretrained(
+        model_name,
+        device_map="auto",
+        load_in_4bit=True,
+        torch_dtype=torch.float16,
+    )
+
+    if tokenizer.pad_token is None:
+        tokenizer.pad_token = tokenizer.eos_token
+
+    return HuggingFacePipeline(pipeline=prepare_pipeline(model, tokenizer))
